単語詳細ページ　初期表示時の関数を別ファイルに分離
